Code refactored for Employee Producer & Consumer
