+++ KafkaCustomSerDe/pom.xml
@@ -56,7 +56,7 @@
                         <!--<configuration>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.bigdata.kafka.EmployeeProducer</mainClass>
+                                    <mainClass>com.bigdata.kafka.producer.EmployeeProducer</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>-->
